Solved the architectural issues:
Chosen the solution idea:

Drop the BackupBackendFactory as there is no sense using it when the Playbook
classes can only work with a single BackupBackend. For instance the ResticPlaybook
and the related components cannot use a BackupBackend other than Restic.
The BackupBackend creation shall be done during the creation of the PlaybookParser objects.
This means that the Playbook and the related classes shall not be generic anymore via the BackupBackendT.